Updated master, more complex average
First update in forever as I am getting back into it. Changed how the average "hue" is calculated, now takes the highest 3 frequencies, assigns them colors, and blends them to get the result. A bit more obtuse but it results in much more dynamic colors. Also fixed the fact it was trying to take 8 frequencies from a 7 frequency EQ. Had to change the structs to have data for the RGB color instead of just HSV hue.